backend/app/modules/models/model.py
from datetime import datetime
from sqlalchemy import Column, Integer, String, DateTime
from extensions import db
class Module(db.Model):
    __tablename__ = 'modules'

    id = Column(Integer, primary_key=True, autoincrement=True)
    name = Column(String(255), nullable=False)
    branch = Column(String(255))
    created_at = Column(DateTime, default=datetime.utcnow)
    updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

    def __repr__(self):
        return f'<Module {self.name}>'
