backend/requirements.txt
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kombu==5.5.2
Mako==1.3.9
MarkupSafe==3.0.2
openai==1.70.0
packaging==24.2
prompt_toolkit==3.0.50
pydantic==2.11.1
pydantic_core==2.33.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
